Change type of ids from Int -> Long

+++ src/android/app/src/main/java/io/github/alexdenton/interviewd/candidate/CandidateFormFragment.kt
@@ -39,7 +39,7 @@ class CandidateFormFragment : RxAwareFragment() {
         vm.withKodein(LazyKodein(appKodein))
 
         if(arguments?.containsKey("candidateId") == true){
-            vm.editing(arguments.getInt("candidateId"))
+            vm.editing(arguments.getLong("candidateId"))
             vm.fetchCandidate()
                     .subscribe {candidate -> setCandidate(candidate) }
                     .lifecycleAware()
