+++ src/android/app/src/online/java/io/github/alexdenton/interviewd/retrofit/InterviewdApiRetrofit.kt
@@ -32,45 +32,45 @@ interface InterviewdApiRetrofit {
     fun getTemplates(): Single<List<TemplateDto>>
 
     @GET("/templates/{id}")
-    fun getTemplate(@Path("id") id: Int): Single<TemplateDto>
+    fun getTemplate(@Path("id") id: Long): Single<TemplateDto>
 
     @POST("/templates")
     fun createTemplate(@Body template: TemplateDto): Single<TemplateDto>
 
     @DELETE("/templates/{id}")
-    fun deleteTemplate(@Path("id") id: Int): Single<TemplateDto>
+    fun deleteTemplate(@Path("id") id: Long): Single<TemplateDto>
 
     @PATCH("/templates/{id}")
-    fun patchTemplate(@Path("id") id: Int, @Body patch: TemplateDto): Single<TemplateDto>
+    fun patchTemplate(@Path("id") id: Long, @Body patch: TemplateDto): Single<TemplateDto>
 
     @GET("/candidates")
     fun getCandidates(): Single<List<CandidateDto>>
 
     @GET("/candidates/{id}")
-    fun getCandidate(@Path("id") id: Int): Single<CandidateDto>
+    fun getCandidate(@Path("id") id: Long): Single<CandidateDto>
 
     @POST("/candidates")
     fun createCandidate(@Body candidate: CandidateDto): Single<CandidateDto>
 
     @DELETE("/candidates")
-    fun deleteCandidate(@Path("id") id: Int): Single<CandidateDto>
+    fun deleteCandidate(@Path("id") id: Long): Single<CandidateDto>
 
     @PATCH("/candidates/{id}")
-    fun patchCandidate(@Path("id") id: Int, @Body patch: CandidateDto): Single<CandidateDto>
+    fun patchCandidate(@Path("id") id: Long, @Body patch: CandidateDto): Single<CandidateDto>
 
     @GET("/interviews")
     fun getInterviews(): Single<List<InterviewDto>>
 
     @GET("/interviews/{id}")
-    fun getInterview(@Path("id") id: Int): Single<InterviewDto>
+    fun getInterview(@Path("id") id: Long): Single<InterviewDto>
 
     @POST("/interviews")
     fun createInterview(@Body interview: InterviewDto): Single<InterviewDto>
 
     @DELETE("/interviews/{id}")
-    fun deleteInterview(@Path("id") id: Int): Single<InterviewDto>
+    fun deleteInterview(@Path("id") id: Long): Single<InterviewDto>
 
     @PATCH("/interviews/{id}")
-    fun patchInterview(@Path("id") id: Int, @Body patch: InterviewDto): Single<InterviewDto>
+    fun patchInterview(@Path("id") id: Long, @Body patch: InterviewDto): Single<InterviewDto>
 
 }
