+++ src/android/app/src/online/java/io/github/alexdenton/interviewd/retrofit/repositories/InterviewRetrofitRepository.kt
@@ -1,16 +1,15 @@
 package io.github.alexdenton.interviewd.retrofit.repositories
 
 import android.content.Context
-import io.github.alexdenton.interviewd.api.InterviewdApi
-import io.github.alexdenton.interviewd.retrofit.RetrofitFactory
-import io.github.alexdenton.interviewd.retrofit.dto.CandidateDto
-import io.github.alexdenton.interviewd.retrofit.dto.InterviewDto
-import io.github.alexdenton.interviewd.retrofit.dto.QuestionDto
 import io.github.alexdenton.interviewd.api.repositories.InterviewRepository
 import io.github.alexdenton.interviewd.entities.Candidate
 import io.github.alexdenton.interviewd.entities.Interview
 import io.github.alexdenton.interviewd.entities.Question
 import io.github.alexdenton.interviewd.retrofit.InterviewdApiRetrofit
+import io.github.alexdenton.interviewd.retrofit.RetrofitFactory
+import io.github.alexdenton.interviewd.retrofit.dto.CandidateDto
+import io.github.alexdenton.interviewd.retrofit.dto.InterviewDto
+import io.github.alexdenton.interviewd.retrofit.dto.QuestionDto
 import io.reactivex.Single
 
 /**
@@ -20,7 +19,7 @@ class InterviewRetrofitRepository(val context: Context) : InterviewRepository {
 
     val client: InterviewdApiRetrofit = RetrofitFactory(context).create(RetrofitFactory.Mode.Local)
 
-    override fun getInterview(id: Int): Single<Interview>
+    override fun getInterview(id: Long): Single<Interview>
             = client.getInterview(id)
             .map { it.toInterview() }
 
@@ -36,7 +35,7 @@ class InterviewRetrofitRepository(val context: Context) : InterviewRepository {
             = client.patchInterview(interview.id, interview.toDto())
             .map { it.toInterview() }
 
-    override fun deleteInterview(id: Int): Single<Interview>
+    override fun deleteInterview(id: Long): Single<Interview>
             = client.deleteInterview(id)
             .map { it.toInterview() }
